pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pydantic-settings = "^2.3.4"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.31" }
loguru = "^0.7.2"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
uvicorn-worker = "^0.2.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.3"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
